requirements.txt
requests
azure-identity
azure-keyvault-secrets
azure-data-tables
mkdocs
